Guided Transfer: Complete purchase notice
This adds a notice to the export page of site settings prompting the user to complete a Guided Transfer purchase. This should
appear when the user has entered host details, but hasn't yet completed checkout.

Change 'Purchase' button text to 'Continue'

Connects notice to Redux

extra JSDoc

Connect in progress card to Redux